Consolidate types; remove Model/types
